requirements.txt
mouseinfo==0.1.3
pillow==12.1.0
pyautogui==0.9.54
pygetwindow==0.0.9
pymsgbox==2.0.1
pyperclip==1.11.0
pypiwin32==223
pyrect==0.2.0
pyscreeze==1.0.1
python-dotenv==1.2.1
pytweening==1.2.0
pywin32==311
